reorganize client code, export types and test helpers
