+++ packages/terafoundation_kafka_connector/src/test-helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './config.js';
+export * from './helpers.js';
