+++ packages/terafoundation_kafka_connector/src/utils.ts
